logqueue-fifo: fix flow-control when fetch-limit() is higher than 65536
In high-performance use cases, users may configure log-iw-size() and
fetch-limit() to be higher than 2^16, which caused flow-control issues,
such as messages stuck in the queue forever.

+++ lib/logqueue-fifo.c
@@ -75,8 +75,8 @@ typedef struct _InputQueue
 {
   struct iv_list_head items;
   WorkerBatchCallback cb;
-  guint16 len;
-  guint16 non_flow_controlled_len;
+  guint32 len;
+  guint32 non_flow_controlled_len;
   guint16 finish_cb_registered;
 } InputQueue;
